+++ backend/routes/artist.ts
@@ -3,9 +3,10 @@ import mongoose from 'mongoose';
 import Artist from '../models/artist';
 import {imagesUpload} from '../multer';
 import {ArtistFromDB, ArtistWithoutId} from '../types';
-import auth from '../middleware/auth';
+import auth, {Auth} from '../middleware/auth';
 import permit from '../middleware/permit';
 import {ObjectId} from 'mongodb';
+import Album from '../models/album';
 
 const artistRouter = express.Router();
 
@@ -69,7 +70,7 @@ artistRouter.patch('/:id/togglePublished', auth, permit(['admin']), async (req, 
   }
 });
 
-artistRouter.delete('/:id', auth, permit(['admin']), async (req, res, next) => {
+artistRouter.delete('/:id', auth, async (req: Auth, res, next) => {
   try {
     const {id} = req.params;
     let _id: ObjectId;
@@ -81,6 +82,13 @@ artistRouter.delete('/:id', auth, permit(['admin']), async (req, res, next) => {
 
     const targetArtist = await Artist.findById(_id);
     if (!targetArtist) return res.status(400).send({error: 'There is no artist to delete'});
+
+    if (req.user?._id.toString() === targetArtist.user.toString() && req.user?.role === 'user') {
+      await Album.deleteOne(_id);
+      return res.send({success: 'Artist has been deleted.'});
+    }
+
+    if (req.user?.role !== 'admin') return res.status(403).send({error: 'Not authorized'});
 
     await Artist.deleteOne(_id);
