+++ backend/routes/track.ts
@@ -4,7 +4,7 @@ import mongoose from 'mongoose';
 import {ObjectId} from 'mongodb';
 import Track from '../models/track';
 import Album from '../models/album';
-import auth from '../middleware/auth';
+import auth, {Auth} from '../middleware/auth';
 import permit from '../middleware/permit';
 import track from '../models/track';
 
@@ -116,7 +116,7 @@ trackRouter.patch('/:id/togglePublished', auth, permit(['admin']), async (req, r
   }
 });
 
-trackRouter.delete('/:id', auth, permit(['admin']), async (req, res, next) => {
+trackRouter.delete('/:id', auth, async (req: Auth, res, next) => {
   try {
     const {id} = req.params;
     let _id: ObjectId;
@@ -127,7 +127,14 @@ trackRouter.delete('/:id', auth, permit(['admin']), async (req, res, next) => {
     }
 
     const targetTrack = await Track.findById(_id);
-    if (!targetTrack) return res.status(400).send({error: 'There is no track to delete'});
+    if (!targetTrack) return res.status(400).send({error: 'There is no track to delete'})
+
+    if (req.user?._id.toString() === targetTrack.user.toString() && req.user?.role === 'user') {
+      await Album.deleteOne(_id);
+      return res.send({success: 'Track has been deleted.'});
+    }
+
+    if (req.user?.role !== 'admin') return res.status(403).send({error: 'Not authorized'});
 
     await track.deleteOne(_id);
